fix warning and doc anomoly on win32 'bout SOCKET
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6714 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/NetHandler.cxx
@@ -22,7 +22,7 @@ bool NetHandler::initHandlers(struct sockaddr_in addr) {
   // udp socket
   int n;
   // we open a udp socket on the same port if alsoUDP
-  if ((udpSocket = socket(AF_INET, SOCK_DGRAM, 0)) < 0) {
+  if ((udpSocket = (int) socket(AF_INET, SOCK_DGRAM, 0)) < 0) {
       nerror("couldn't make udp connect socket");
       return false;
   }
